pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glove80-keymap-visualizer"
version = "0.1.0"
description = "Generate PDF visualizations of Glove80 keyboard layers from ZMK keymap files"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dave Sifry"}
]
keywords = ["glove80", "zmk", "keyboard", "keymap", "visualization", "pdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "keymap-drawer>=0.18.0",
    "cairosvg>=2.7.0",
    "PyPDF2>=3.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
glove80-viz = "glove80_visualizer.cli:main"

[project.urls]
Homepage = "https://github.com/dsifry/glove80-keymap-visualizer"
Repository = "https://github.com/dsifry/glove80-keymap-visualizer"
Issues = "https://github.com/dsifry/glove80-keymap-visualizer/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/glove80_visualizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src/glove80_visualizer --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
